syrk benchmark parameters in a separate header
