+++ include/bench/Benchmark.hpp
@@ -8,7 +8,6 @@
 
 #include <benchmark/benchmark.h>
 
-#define BENCHMARK_MAX_SYRK 50
 #define BENCHMARK_MAX_POTRF 50
 
 
@@ -16,7 +15,5 @@ namespace blast :: benchmark
 {
     using namespace ::benchmark;
 
-
-    void syrkBenchArguments(internal::Benchmark * b);
     void trmmBenchArguments(internal::Benchmark * b);
 }
