+++ include/bench/Complexity.hpp
@@ -67,16 +67,6 @@ namespace blast :: benchmark
     }
 
 
-    /// @brief Algorithmic complexity of syrk
-    inline Complexity complexitySyrk(std::size_t m, std::size_t k)
-    {
-        return {
-            {"add", m * (m + 1) * k / 2},
-            {"mul", m * (m + 1) * k / 2}
-        };
-    }
-
-
     template <typename Map>
     void setCounters(Map& counters, Complexity const& c)
     {
